Dynamically discover SAMPLE_DATA shared db name
There is no official standard for the name of the SAMPLE DATA shared db in a Snowflake account (it could be SNOWFLAKE_SAMPLE_DATA, SF_SAMPLE_DATA, SAMPLE_DATA, etc). This fix utilizes the "SHOW SHARES" command with a wildcard pattern to auto-discover the shared db name. Also, updated one of the "sql.fmt" files to have a placeholder for the shared database name ({SNOWFLAKE_SAMPLE_DATA}). This required the addition of a new optional parameter in the "read_queries()" method.